Large number of changes to fix the waypoint tracking code. The problem was that the settings for triggering a change to a new track of waypoints wasn't updated to the new 0..1000 range for the RC inputs and so was triggering after a little while. Updating those values to the proper ones in that range fixed the problem (200 and 800).
Modified the groundstation and code_gen to display the current and next waypoints. This included upping the baud rate from 57600 to 115200 for UART1.

Waypoints are now stored as uint32s instead of doubles. There was no reason to keep them as floating point, so I changed them. I figure that a meter is a high enough resolution for the boat considering its size.

The calibration triggers from the RC input now account for transitions to disconnected by ignoring them.