Found a glitch in RC reception code where the low heartbeat pulses on the throttle channel weren't handled as 0 values and were instead handled as valid throttle commands. This is now fixed so that the heartbeat values are now interpreted as 0 and so they're dealt with properly now. This code applies to code_gen and RC_transceiver_test.mdl.
I have added an additional model based off of code_gen that reduces the possible commands for the rudder and throttle. The throttle commands are binned heavily into about 15 different steps. The rudder values are binned into about 9 or so and they can only be stepped between incrementally instead of rounding the continuous input. This will allow for more precise measurements of turn rate as a function of speed and rudder angle.