Transmit all parameter values when the vehicle is set to autonomous mode.
This entailed a heavy refactoring of the code that dealt with the mode of the vehicle.
Code was duplicated and not architected nicely. I've since moved all the mode logic into
the PrimaryNode library and the ParameterHelper and MavlinkGlue libs now reference that
logic for changing the mode. This means alterations to events that are triggered by a
mode change only need to be in one place.