Make the RTB state a reset error and have a clear enable/disable path for it.
The RTB flag is therefore a way to disable autonomous mode once an RTB criteria is met,
but then even if those criteria go away, have it stay in reset. I want it to be this way
so that I have to switch it to manual mode, check everything, and then re-enable autonomous
mode.